+++ docs/changelog.md
@@ -1,6 +1,15 @@
 Terrific Audio Driver Changelog
 ===============================
 
+Version 0.0.11
+==============
+
+Bytecode changes:
+ * The loop instructions now use a traditional stack.
+    * The maximum number of nested loops has increased.
+
+
+
 Version 0.0.10
 ==============